+++ js/firesight/CalcGrid.js
@@ -39,10 +39,12 @@ var Grid = require("../../www/js/shared/Grid");
             if (rects && rects.length > 0) {
                 var grid = (rects && rects.length > 4) ? Grid.createFromPoints(rects) : null;
                 if (grid) {
+                    var stats = grid.statsFromPoints(rects);
                     var result = {
                         origin: grid.origin,
                         angle: grid.angle,
                         cellSize: grid.cellSize,
+                        rmse: {x: stats.xRMSE, y:stats.yRMSE},
                     }
                     console.log("INFO\t: CalcGrid " + jsonDstPath + ") ");
                     firesight.verbose && console.log("DEBUG\t: " + JSON.stringify(result));